+++ packages/core/src/utils/sorter/ComponentSorter.ts
@@ -3,7 +3,8 @@ import { CanvasSpotBuiltInTypes } from "../../canvas/model/CanvasSpot";
 import Component from "../../dom_components/model/Component";
 import EditorModel from "../../editor/model/Editor";
 import { ComponentNode } from "./ComponentNode";
-import Sorter, { SorterContainerContext, PositionOptions, SorterDragBehaviorOptions, SorterEventHandlers } from "./Sorter";
+import Sorter from "./Sorter";
+import { SorterContainerContext, PositionOptions, SorterDragBehaviorOptions, SorterEventHandlers } from './types';
 
 const targetSpotType = CanvasSpotBuiltInTypes.Target;
 const spotTarget = {
