+++ packages/core/src/utils/sorter/Sorter.ts
@@ -1,5 +1,4 @@
 import { bindAll, isFunction } from 'underscore';
-import CanvasModule from '../../canvas';
 import { $, View } from '../../common';
 import EditorModel from '../../editor/model/Editor';
 import { off, on } from '../dom';
@@ -7,76 +6,8 @@ import { TreeSorterBase } from './TreeSorterBase';
 import { DropLocationDeterminer } from './DropLocationDeterminer';
 import { PlaceholderClass } from './PlaceholderClass';
 import { getMergedOptions, getDocuments, matches, closest } from './SorterUtils';
-
-export interface Dimension {
-  top: number;
-  left: number;
-  height: number;
-  width: number;
-  offsets: ReturnType<CanvasModule['getElementOffsets']>;
-  dir?: boolean;
-  el?: HTMLElement;
-  indexEl?: number;
-}
-
-export interface Position {
-  index: number;
-  indexEl: number;
-  method: string;
-}
-
-export enum SorterDirection {
-  Vertical = "Vertical",
-  Horizontal = "Horizontal",
-  BothDirections = "BothDirections"
-}
-
-export interface SorterContainerContext {
-  container: HTMLElement;
-  containerSel: string;
-  itemSel: string;
-  pfx: string;
-  document: Document;
-  placeholderElement: HTMLElement;
-  customTarget?: Function;
-}
-
-export interface PositionOptions {
-  windowMargin?: number;
-  borderOffset?: number;
-  offsetTop?: number;
-  offsetLeft?: number;
-  canvasRelative?: boolean;
-  scale?: number;
-  relative?: boolean;
-}
-
-export interface SorterEventHandlers<T> {
-  onStartSort?: (sourceNode: TreeSorterBase<T>, container?: HTMLElement) => void;
-  onDragStart?: (mouseEvent: MouseEvent) => void;
-  onMouseMove?: Function;
-  onDrop?: (targetNode: TreeSorterBase<T>, sourceNode: TreeSorterBase<T>, index: number) => void;
-  onTargetChange?: (oldTargetNode: TreeSorterBase<T>, newTargetNode: TreeSorterBase<T>) => void;
-  onPlaceholderPositionChange?: (dims: Dimension[], newPosition: Position) => void;
-  onEndMove?: Function;
-}
-
-export interface SorterDragBehaviorOptions {
-  dragDirection: SorterDirection;
-  ignoreViewChildren?: boolean;
-  nested?: boolean;
-  selectOnEnd?: boolean;
-}
-
-export interface SorterOptions<T> {
-  em: EditorModel;
-  treeClass: new (model: T) => TreeSorterBase<T>;
-
-  containerContext: SorterContainerContext;
-  positionOptions: PositionOptions;
-  dragBehavior: SorterDragBehaviorOptions;
-  eventHandlers?: SorterEventHandlers<T>;
-}
+import { SorterContainerContext, PositionOptions, SorterDragBehaviorOptions, SorterEventHandlers, Dimension, Position } from './types';
+import { SorterOptions } from './types';
 
 export type RequiredEmAndTreeClassPartialSorterOptions<T> = Partial<SorterOptions<T>> & {
   em: EditorModel;
