+++ packages/core/src/utils/sorter/StyleManagerSorter.ts
@@ -2,7 +2,8 @@ import EditorModel from "../../editor/model/Editor";
 import Layer from "../../style_manager/model/Layer";
 import Layers from "../../style_manager/model/Layers";
 import { LayerNode } from "./LayerNode";
-import Sorter, { SorterContainerContext, PositionOptions, SorterDragBehaviorOptions, SorterEventHandlers } from "./Sorter";
+import Sorter from "./Sorter";
+import { SorterContainerContext, PositionOptions, SorterDragBehaviorOptions, SorterEventHandlers } from './types';
 
 export default class StyleManagerSorter extends Sorter<Layers | Layer> {
     constructor({
